First steps toward de-virtualizing TTransport.
- Pull readAll out of TTransport to be a global templatized helper function.
- Create a TVirtualTransport as a bridge between the virtual TTransport
  and nonvirtual implementations.

Chad Walters wrote the first version of this patch.  I (David Reiss) just
pulled out the simplest parts.  I left the TTransportUtils part out for now.
The calls to the underlying transports are going to be relatively rare, and
probably quite expensive (a system call), so de-virtualizing them will probably
not be that big of a performance win.