Prevent TTransport <-> TVirtualTransport infinite loops.
Just like the TPipedTransport issue, we were getting infinite loops in
TBufferBase.  Specifically:
  -               Caller
  calls           TTransport::flush()
  calls           TTransport::flush_virt()
  virtually is    TVirtualTransport::flush_virt()
  calls           TBufferBase::flush()
  nonvirtually is TTransport::flush()
Fixed by copying all of the nonvirtual stubs into TVirtualTransport,
so TBufferBase::flush() nonvirutally is TVirtualTransport::flush(),
which is a stub.  We should try to remove the stubs from TTransport
ASAP, but that is kind of a pain because of TPipedTransport and
TFileTransport and their virtual inheritance.