THRIFT-338: Beginnings of a compact protocol implementation in Erlang.
This is not tested yet. The zigzag encoding and varint stuff has been
tested in the REPL, but none of the actual protocol stuff.

TODO: we need to add factory hooks into thrift_server and/or thrift_socket_server
to override the protocol implementation as an option. Then we need to write some
tests that make sure it actually works.